openai_requirements.txt
websocket-client==1.7.0
websockets==15.0.1
python-dotenv==1.0.1